Update facades to drop `Lib` suffix, delete std-facade since it has nothing useful
