--- facades/std/build.json
@@ -1,4 +1,0 @@
-{
-  "typings": [],
-  "dependencies": []
-}
