--- facades/std/build.sbt
@@ -1,13 +1,0 @@
-organization := "org.scalablytyped"
-name := "std-facade"
-version := "201907081111-532dd6"
-scalaVersion := "2.12.8"
-enablePlugins(ScalaJSPlugin)
-libraryDependencies ++= Seq(
-  "com.olvind" %%% "scalablytyped-runtime" % "2.1.0")
-publishArtifact in packageDoc := false
-scalacOptions ++= List("-P:scalajs:sjsDefinedByDefault", "-g:notailcalls")
-licenses += ("MIT", url("http://opensource.org/licenses/MIT"))
-bintrayRepository := "ScalablyTyped"
-resolvers += Resolver.bintrayRepo("oyvindberg", "ScalablyTyped")
-
