--- facades/std/project/plugins.sbt
@@ -1,2 +1,0 @@
-addSbtPlugin("org.scala-js" %% "sbt-scalajs" % "0.6.28")
-addSbtPlugin("org.foundweekends" % "sbt-bintray" % "0.5.4")
