--- facades/std/readme.md
@@ -1,15 +1,0 @@
-
-# Scala.js typings for std-facade
-
-
-
-
-## Note
-This library has been generated from typescript code from first party type definitions.
-
-Provided with :purple_heart: from [ScalablyTyped](https://github.com/oyvindberg/ScalablyTyped)
-
-## Usage
-See [the main readme](../../readme.md) for instructions.
-
-
